:recycle: HitboxHitResult 에 서로 때린 결과가 중복되지 않도록 수정
* HitboxHitResult 클래스에 IEquatable 인터페이스와 GetHashCode() 함수 구현
* HitboxResolverComponent 가 HashSet<HitboxHitResult> 를 사용하도록 변경
