+++ Runelight-Smash-2021/Assets/Scenes/SampleScene.unity
@@ -1095,6 +1095,55 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 13
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &424674845
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 424674846}
+  - component: {fileID: 424674847}
+  m_Layer: 0
+  m_Name: Attack
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &424674846
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 424674845}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0.3, y: 0.1, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 2025968842}
+  m_RootOrder: 4
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &424674847
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 424674845}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 66844a1ee6df8414d8e6e75fd2b33fb8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  hitRateFrame: 60
+  hitboxDurationFrame: 5
+  hitboxInfo:
+    id: 0
+    groupId: 0
+    type: 1
 --- !u!1 &441413159
 GameObject:
   m_ObjectHideFlags: 0
@@ -3021,6 +3070,10 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   hitRateFrame: 60
   hitboxDurationFrame: 5
+  hitboxInfo:
+    id: 0
+    groupId: 0
+    type: 1
 --- !u!4 &882533424
 Transform:
   m_ObjectHideFlags: 0
@@ -4826,6 +4879,7 @@ Transform:
   - {fileID: 418088295}
   - {fileID: 442199920}
   - {fileID: 1295656624}
+  - {fileID: 424674846}
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
