+++ Runelight-Smash-2021/Assets/Scripts/Hitbox/HitboxHitResult.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2d0732703b990ce43a389caa5deb5e9c
+guid: 391a807142a10ec46abc7fbff2394da5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
